move manifest loading off to context too
