+++ cmd/apply.go
@@ -4,6 +4,7 @@ import (
 	"io"
 	"os"
 
+	"github.com/gomicro/concord/manifest"
 	"github.com/gomicro/concord/report"
 	"github.com/spf13/cobra"
 )
@@ -30,26 +31,22 @@ func NewApplyCmd(out io.Writer) *cobra.Command {
 
 func applyRun(cmd *cobra.Command, args []string) error {
 	file := args[0]
-
-	org, err := readManifest(file)
-	if err != nil {
-		return handleError(cmd, err)
-	}
+	cmd.SetContext(manifest.WithManifest(cmd.Context(), file))
 
 	report.PrintHeader("Org")
 	report.Println()
 
-	err = membersRun(cmd, args, org, false)
+	err := membersRun(cmd, args, false)
 	if err != nil {
 		return handleError(cmd, err)
 	}
 
-	err = teamsRun(cmd, args, org, false)
+	err = teamsRun(cmd, args, false)
 	if err != nil {
 		return handleError(cmd, err)
 	}
 
-	err = reposRun(cmd, args, org, false)
+	err = reposRun(cmd, args, false)
 	if err != nil {
 		return handleError(cmd, err)
 	}
