+++ cmd/apply_teams.go
@@ -4,6 +4,7 @@ import (
 	"io"
 	"os"
 
+	"github.com/gomicro/concord/manifest"
 	"github.com/gomicro/concord/report"
 	"github.com/spf13/cobra"
 )
@@ -28,14 +29,10 @@ func NewApplyTeamsCmd(out io.Writer) *cobra.Command {
 
 func applyTeamsRun(cmd *cobra.Command, args []string) error {
 	file := args[0]
-
-	org, err := readManifest(file)
-	if err != nil {
-		return handleError(cmd, err)
-	}
+	cmd.SetContext(manifest.WithManifest(cmd.Context(), file))
 
 	report.PrintHeader("Org")
 	report.Println()
 
-	return teamsRun(cmd, args, org, false)
+	return teamsRun(cmd, args, false)
 }
