+++ cmd/check_teams.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/gomicro/concord/client"
 	gh_pb "github.com/gomicro/concord/github/v1"
+	"github.com/gomicro/concord/manifest"
 	"github.com/gomicro/concord/report"
 	"github.com/google/go-github/v56/github"
 	"github.com/spf13/cobra"
@@ -34,20 +35,21 @@ func NewCheckTeamsCmd(out io.Writer) *cobra.Command {
 
 func checkTeamsRun(cmd *cobra.Command, args []string) error {
 	file := args[0]
-
-	org, err := readManifest(file)
-	if err != nil {
-		return handleError(cmd, err)
-	}
+	cmd.SetContext(manifest.WithManifest(cmd.Context(), file))
 
 	report.PrintHeader("Org")
 	report.Println()
 
-	return teamsRun(cmd, args, org, true)
+	return teamsRun(cmd, args, true)
 }
 
-func teamsRun(cmd *cobra.Command, args []string, org *gh_pb.Organization, dry bool) error {
+func teamsRun(cmd *cobra.Command, args []string, dry bool) error {
 	ctx := cmd.Context()
+
+	org, err := manifest.OrgFromContext(ctx)
+	if err != nil {
+		return handleError(cmd, err)
+	}
 
 	clt, err := client.ClientFromContext(ctx)
 	if err != nil {
